Initial version
This program arose from a joke to make a permutation calculator. The program expects two permutations of positive integers in cycle notation as inputs. The program outputs the result of the composition of the input permutations. Two different algorithms are used to compute the output. Logging the execution time of the algorithms.